requirements.txt
pytest
requests
jsonschema
selene==2.0.0rc9
allure-pytest